scripts/requirements.txt
GitPython >= 1.7.0
jira >= 3.6.0